DRTII-1346 Refactor persistence
