+++ jvm/src/main/scala/uk/gov/homeoffice/drt/prediction/persistence/Flight.scala
@@ -3,7 +3,7 @@ package uk.gov.homeoffice.drt.prediction.persistence
 import akka.actor.ActorSystem
 import akka.util.Timeout
 import uk.gov.homeoffice.drt.prediction.category.FlightCategory
-import uk.gov.homeoffice.drt.prediction.{ModelCategory, PersistenceImpl}
+import uk.gov.homeoffice.drt.prediction.{ModelCategory, ActorModelPersistenceImpl}
 import uk.gov.homeoffice.drt.time.{SDate, SDateLike}
 
 import scala.concurrent.ExecutionContext
@@ -13,7 +13,7 @@ case class Flight()
                   val ec: ExecutionContext,
                   val timeout: Timeout,
                   val system: ActorSystem
-                 ) extends PersistenceImpl {
+                 ) extends ActorModelPersistenceImpl {
   override val now: () => SDateLike = () => SDate.now()
   override val modelCategory: ModelCategory = FlightCategory
 }
